backend/telegram_bot/requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
httpx==0.27.0
redis==5.0.1